requirements.txt
selenium==4.39.0
pytest-selenium
flake8
autopep8
pytest-order
pytest-html
